# src/main/java/com/dealerengine/www/api/DealersApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (6.0.0-beta).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.dealerengine.www.api;

import com.dealerengine.www.model.Dealer;
import com.dealerengine.www.model.Error;
import com.dealerengine.www.model.Information;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.multipart.MultipartFile;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
import java.util.Map;
import javax.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2022-05-20T16:17:24.620-05:00[America/Chicago]")
@Validated
@Tag(name = "dealers", description = "the dealers API")
public interface DealersApi {

    default DealersApiDelegate getDelegate() {
        return new DealersApiDelegate() {};
    }

    /**
     * POST /dealers/addNew : Add New Dealer
     * To add a new dealer
     *
     * @param dealer Dealer Details (optional)
     * @return Dealer added successfully (status code 200)
     *         or Error in adding dealer (status code 400)
     *         or User not authorized for adding dealer (status code 401)
     */
    @Operation(
        operationId = "addNewDealer",
        summary = "Add New Dealer",
        responses = {
            @ApiResponse(responseCode = "200", description = "Dealer added successfully", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Dealer.class))
            }),
            @ApiResponse(responseCode = "400", description = "Error in adding dealer", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Error.class))
            }),
            @ApiResponse(responseCode = "401", description = "User not authorized for adding dealer", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Error.class))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/dealers/addNew",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    default ResponseEntity<Dealer> addNewDealer(
        @Parameter(name = "Dealer", description = "Dealer Details") @Valid @RequestBody(required = false) Dealer dealer
    ) {
        return getDelegate().addNewDealer(dealer);
    }


    /**
     * DELETE /dealers/{dealerId} : Delete Existing Dealer
     * To delete an existing dealer
     *
     * @param dealerId Unique id of the dealer (required)
     * @return Dealer deleted successfully (status code 200)
     *         or User not authorized for deleting dealer (status code 401)
     *         or Dealer not found (status code 404)
     */
    @Operation(
        operationId = "deleteExistingDealer",
        summary = "Delete Existing Dealer",
        responses = {
            @ApiResponse(responseCode = "200", description = "Dealer deleted successfully", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Information.class))
            }),
            @ApiResponse(responseCode = "401", description = "User not authorized for deleting dealer", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Error.class))
            }),
            @ApiResponse(responseCode = "404", description = "Dealer not found", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Error.class))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = "/dealers/{dealerId}",
        produces = { "application/json" }
    )
    default ResponseEntity<Information> deleteExistingDealer(
        @Parameter(name = "dealerId", description = "Unique id of the dealer", required = true) @PathVariable("dealerId") Integer dealerId
    ) {
        return getDelegate().deleteExistingDealer(dealerId);
    }


    /**
     * GET /dealers/{dealerId} : Get Existing Dealer
     * To get an existing dealer
     *
     * @param dealerId Unique id of the dealer (required)
     * @return Dealer retrieved successfully (status code 200)
     *         or Dealer not found (status code 404)
     */
    @Operation(
        operationId = "getExistingDealer",
        summary = "Get Existing Dealer",
        responses = {
            @ApiResponse(responseCode = "200", description = "Dealer retrieved successfully", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Dealer.class))
            }),
            @ApiResponse(responseCode = "404", description = "Dealer not found", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Error.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/dealers/{dealerId}",
        produces = { "application/json" }
    )
    default ResponseEntity<Dealer> getExistingDealer(
        @Parameter(name = "dealerId", description = "Unique id of the dealer", required = true) @PathVariable("dealerId") Integer dealerId
    ) {
        return getDelegate().getExistingDealer(dealerId);
    }


    /**
     * PUT /dealers/{dealerId} : Update Existing Dealer
     * To update an existing dealer
     *
     * @param dealerId Unique id of the dealer (required)
     * @param dealer  (optional)
     * @return Dealer updated successfully (status code 200)
     *         or Error in updating dealer (status code 400)
     *         or User not authorized for updating dealer (status code 401)
     *         or Dealer not found (status code 404)
     */
    @Operation(
        operationId = "updateExistingDealer",
        summary = "Update Existing Dealer",
        responses = {
            @ApiResponse(responseCode = "200", description = "Dealer updated successfully", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Dealer.class))
            }),
            @ApiResponse(responseCode = "400", description = "Error in updating dealer", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Error.class))
            }),
            @ApiResponse(responseCode = "401", description = "User not authorized for updating dealer", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Error.class))
            }),
            @ApiResponse(responseCode = "404", description = "Dealer not found", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Error.class))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = "/dealers/{dealerId}",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    default ResponseEntity<Dealer> updateExistingDealer(
        @Parameter(name = "dealerId", description = "Unique id of the dealer", required = true) @PathVariable("dealerId") Integer dealerId,
        @Parameter(name = "Dealer", description = "") @Valid @RequestBody(required = false) Dealer dealer
    ) {
        return getDelegate().updateExistingDealer(dealerId, dealer);
    }

}