# src/main/java/com/dealerengine/www/api/UsersApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (6.0.0-beta).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.dealerengine.www.api;

import com.dealerengine.www.model.AdminUser;
import com.dealerengine.www.model.Error;
import com.dealerengine.www.model.SignInRequest;
import com.dealerengine.www.model.Token;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.media.ExampleObject;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.multipart.MultipartFile;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
import java.util.Map;
import javax.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2022-05-20T16:17:24.620-05:00[America/Chicago]")
@Validated
@Tag(name = "users", description = "the users API")
public interface UsersApi {

	default UsersApiDelegate getDelegate() {
		return new UsersApiDelegate() {
		};
	}

	/**
	 * POST /users/adminUserSignIn : Admin User SignIn To signin an admin user
	 *
	 * @param signInRequest Admin User SignIn Request (optional)
	 * @return Admin user signed in successfully (status code 200) or Authentication
	 *         failed (status code 401)
	 */
	@Operation(operationId = "adminUserSignIn", summary = "Admin User SignIn", requestBody = @io.swagger.v3.oas.annotations.parameters.RequestBody(required = true, content = @Content(schema = @Schema(implementation = AdminUser.class), examples = {
			@ExampleObject(name = "signin_request", description = "SignIn Request", value = "{\n"
					+ "	\"userName\": \"Marva27\",\n" + "\"password\": \"Boeing@27\"\n" + "}") })), responses = {
							@ApiResponse(responseCode = "200", description = "Admin user signed in successfully", content = {
									@Content(mediaType = "application/json", schema = @Schema(implementation = Token.class)) }),
							@ApiResponse(responseCode = "401", description = "Authentication failed", content = {
									@Content(mediaType = "application/json", schema = @Schema(implementation = Error.class)) }) })
	@RequestMapping(method = RequestMethod.POST, value = "/users/adminUserSignIn", produces = {
			"application/json" }, consumes = { "application/json" })
	default ResponseEntity<Token> adminUserSignIn(
			@Parameter(name = "SignInRequest", description = "Admin User SignIn Request") @Valid @RequestBody(required = false) SignInRequest signInRequest) {
		return getDelegate().adminUserSignIn(signInRequest);
	}

	/**
	 * POST /users/adminUserSignUp : Admin User SignUp To signup a new admin user
	 *
	 * @param adminUser Admin User SignUp Request (optional)
	 * @return Admin user created successfully (status code 200) or Error in signing
	 *         up new admin user (status code 400)
	 */
	@Operation(operationId = "adminUserSignUp", summary = "Admin User SignUp", requestBody = @io.swagger.v3.oas.annotations.parameters.RequestBody(required = true, content = @Content(schema = @Schema(implementation = AdminUser.class), examples = {
			@ExampleObject(name = "signup_request", value = "{\n" + "	\"firstName\": \"Srinivasan\",\n"
					+ "\"lastName\": \"Ramu\",\n" + "\"emailAddress\": \"srinimarva@gmail.com\",\n"
					+ "\"userName\": \"Marva27\",\n" + "\"password\": \"Boeing@27\"\n" + "}") })), responses = {
							@ApiResponse(responseCode = "200", description = "Admin user created successfully", content = {
									@Content(mediaType = "application/json", schema = @Schema(implementation = AdminUser.class), examples = {
											@ExampleObject(name = "signup_success", value = "{\n"
													+ "	\"firstName\": \"Srinivasan\",\n" + "\"lastName\": \"Ramu\",\n"
													+ "\"emailAddress\": \"srinimarva@gmail.com\",\n"
													+ "\"userName\": \"Marva27\",\n" + "\"password\": \"Boeing@27\"\n"
													+ "}") }) }),
							@ApiResponse(responseCode = "400", description = "Error in signing up new admin user", content = {
									@Content(mediaType = "application/json", schema = @Schema(implementation = Error.class), examples = {
											@ExampleObject(name = "required_fields_missing", value = "{\n"
													+ "	\"id\": \"badRequest|userName\",\n"
													+ "	\"message\": \"userName is required\"\n" + "}"),
											@ExampleObject(name = "user_already_exists", value = "{\n"
													+ "	\"id\": \"badRequest|userName\",\n"
													+ "	\"message\": \"userName is already taken\"\n" + "}") }) }) })
	@RequestMapping(method = RequestMethod.POST, value = "/users/adminUserSignUp", produces = {
			"application/json" }, consumes = { "application/json" })
	default ResponseEntity<AdminUser> adminUserSignUp(
			@Parameter(name = "AdminUser", description = "Admin User SignUp Request") @Valid @RequestBody(required = false) AdminUser adminUser) {
		return getDelegate().adminUserSignUp(adminUser);
	}

}